Added the option to remove a link
Added a delete button to every element in the main table on the front page.
Removed ids from being visible on the front page.
Updated readme.
Updated methods in db_module.py (create_sql_table and get_price_change) to work properly without the need to count the tables because it caused issues if a table was deleted.